Extend the Response interface and implementations in server package
The `Response` interface and its implementations `TransactionResponse` and `RespResponse` have been expanded in this commit. Methods to write double values, maps, null values and booleans were added. Also, the `writeMap()` method is utilised in the `HelloHandler` class. It modifies `GetReadVersionHandler` and corresponding test to write readVersion as double instead of integer.

+++ kronotop/src/main/java/com/kronotop/foundationdb/GetReadVersionHandler.java
@@ -51,6 +51,6 @@ class GetReadVersionHandler implements Handler {
         Attribute<Transaction> transactionAttr = channel.attr(ChannelAttributes.TRANSACTION);
         Transaction tr = transactionAttr.get();
         Long readVersion = tr.getReadVersion().join();
-        response.writeInteger(readVersion);
+        response.writeDouble(readVersion);
     }
 }
